Unify Writable/Redable Stream
